requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
typer>=0.12.0
langgraph>=0.0.15
requests>=2.31.0
python-dotenv>=1.0.1
pyodbc>=5.1.0
sqlalchemy>=2.0.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
